requirement.txt
python-dotenv
livekit-agents
livekit-plugins-openai
livekit-plugins-silero
livekit-plugins-turn-detector
livekit-plugins-noise-cancellation
torch